добавление словарь
